fixes issue 17
issues with getResources(). it was just plain broken.

note, we now return the current directory as a resource (it is noted in the DavResource.isCurrentDirectory()) this maps to the way the webdav protocol works better and saves another hit to the server to get that information if you need it.

also added some unit tests


git-svn-id: http://sardine.googlecode.com/svn/trunk@61 bf195e34-ef88-11de-8eb4-853c26953258